% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basictools.R
\docType{_PACKAGE}
\name{basictools}
\alias{basictools}
\title{basictools: Basic Tools Package}
\description{
A set of basic functions to support your daily work.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aldisrauf/basictools}
  \item Report bugs at \url{https://github.com/aldisrauf/basictools/issues}
}

}
\keyword{internal}
