% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bridge23_kab}
\alias{bridge23_kab}
\title{Mapping of Kabupaten (Municipalities) IDs and Names for 2022 and 2023}
\format{
A data frame with 4 columns:
\describe{
  \item{idkab22}{Numeric. The kabupaten ID for the year 2022.}
  \item{nmkab22}{Character. The kabupaten name for the year 2022.}
  \item{idkab23}{Numeric. The kabupaten ID for the year 2023.}
  \item{nmkab23}{Character. The kabupaten name for the year 2023.}
}
}
\usage{
bridge23_kab
}
\description{
A dataset containing the mappings between kabupaten (regency) IDs and names for the years 2022 and 2023.
}
\examples{
\dontrun{
  data(bridge23_kab)
  head(bridge23_kab)
}
}
\keyword{datasets}
