% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bridge23_prov}
\alias{bridge23_prov}
\title{Mapping of Provinsi (Province) IDs and Names for 2022 and 2023}
\format{
A data frame with 4 columns:
\describe{
  \item{idprov22}{Numeric. The provinsi ID for the year 2022.}
  \item{nmprov22}{Character. The provinsi name for the year 2022.}
  \item{idprov23}{Numeric. The provinsi ID for the year 2023.}
  \item{nmprov23}{Character. The provinsi name for the year 2023.}
}
}
\usage{
bridge23_prov
}
\description{
A dataset containing the mappings between provinsi (province) IDs and names for the years 2022 and 2023.
}
\examples{
\dontrun{
  data(bridge23_prov)
  head(bridge23_prov)
}
}
\keyword{datasets}
