% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Correlation Matrix}
\usage{
cor_matrix(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A correlation matrix
}
\description{
Computes the correlation matrix of all numeric columns in a data frame.
}
\examples{
# cor_matrix(mtcars)
}
