% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_header.R
\name{create_header}
\alias{create_header}
\title{Create a Custom Header for a Data Frame}
\usage{
create_header(column, size)
}
\arguments{
\item{column}{A character vector of column names.}

\item{size}{An integer vector specifying the size (number of columns) for each header.}
}
\value{
A data frame with the specified custom header.
}
\description{
This function generates a custom header for a data frame based on specified column names and sizes.
}
\examples{
create_header(c("Header1", "Header2"), c(2, 3))
}
