% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_header.R
\name{export_with_header}
\alias{export_with_header}
\title{Export Data with Custom Headers to an Excel File}
\usage{
export_with_header(
  wb,
  sheet_name,
  header_list,
  data,
  filename,
  print_df_header = FALSE
)
}
\arguments{
\item{wb}{An openxlsx workbook object.}

\item{sheet_name}{A character string specifying the name of the sheet.}

\item{header_list}{A list of lists, where each sublist contains a character vector of column names and an integer vector of sizes for each header.}

\item{data}{A data frame to be written to the Excel sheet.}

\item{filename}{A character string specifying the name of the output Excel file.}

\item{print_df_header}{A logical value indicating whether to print the data frame headers (default is FALSE).}
}
\value{
None. The function saves the Excel file with the specified name.
}
\description{
This function exports data to an Excel file with custom headers and optional data frame headers.
}
\examples{
wb <- openxlsx::createWorkbook()
openxlsx::addWorksheet(wb, "Sheet1")
header_list <- list(list(c("Header1", "Header2"), c(2, 3)))
data <- data.frame(A = 1:3, B = 4:6, C = 7:9, D = 10:12, E = 13:15)
export_with_header(wb, "Sheet1", header_list, data, "output.xlsx")
}
