% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_e.R
\name{find_e}
\alias{find_e}
\title{Calculate Margin of Error for Estimating a Proportion}
\usage{
find_e(p, N, n, deff = 1, rr = 0.9, alpha = 0.05, moe_type = "relative")
}
\arguments{
\item{p}{Numeric. The estimated proportion (must be between 0 and 1).}

\item{N}{Numeric. The population size (must be greater than 0).}

\item{n}{Numeric. The sample size (must be greater than 0).}

\item{deff}{Numeric. The design effect, default is 1.}

\item{rr}{Numeric. The expected response rate, default is 0.9.}

\item{alpha}{Numeric. The significance level, default is 0.05.}

\item{moe_type}{Character. The type of margin of error, either "relative" or "absolute". Default is "relative".}
}
\value{
The calculated margin of error.
}
\description{
This function calculates the margin of error for estimating a proportion in a finite population, given the population size, sample size, design effect, response rate, confidence level, and type of margin of error (relative or absolute).
}
\examples{
find_e(p = 0.5, N = 10000, n = 400, deff = 1, rr = 0.9, alpha = 0.05, moe_type = "relative")
}
