% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_e.R
\name{find_e_mean}
\alias{find_e_mean}
\title{Calculate Margin of Error for Estimating a Mean}
\usage{
find_e_mean(x, sd, N, n, deff = 1, rr = 0.9, alpha = 0.05)
}
\arguments{
\item{x}{Numeric. The estimated mean.}

\item{sd}{Numeric. The estimated standard deviation.}

\item{N}{Numeric. The population size (must be greater than 0).}

\item{n}{Numeric. The sample size (must be greater than 0).}

\item{deff}{Numeric. The design effect, default is 1.}

\item{rr}{Numeric. The expected response rate, default is 0.9.}

\item{alpha}{Numeric. The significance level, default is 0.05.}
}
\value{
The calculated margin of error.
}
\description{
This function calculates the margin of error for estimating a mean in a finite population, given the population size, sample size, design effect, response rate, confidence level.
}
\examples{
find_e(x = 20, sd = 5, N = 10000, n = 400, deff = 1, rr = 0.9, alpha = 0.05)
}
