% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeMem.R
\name{freeMem}
\alias{freeMem}
\title{Free up memory by removing objects from the R environment}
\usage{
freeMem(objs = NULL, exclude = F, echo = F)
}
\arguments{
\item{objs}{A character vector of the names of objects to remove from the R
environment. If \code{NULL}, all objects in the global environment will be
removed.}

\item{exclude}{A logical value indicating whether to exclude the objects named
in \code{objs} from removal instead of removing them. The default is
\code{FALSE}.}

\item{echo}{A logical value indicating whether to print the amount of memory
freed after removing the objects. The default is \code{FALSE}.}
}
\value{
This function is called for its side effects, and does not return a
  value.
}
\description{
This function can be used to remove one or more objects from the R environment
in order to free up memory. By default, all objects in the global environment
will be removed.
}
\examples{
# Remove all objects from the global environment
freeMem()

# Remove specific objects from the global environment
freeMem(c("myData", "myModel"))

# Exclude specific objects from removal
freeMem(c("myData", "myModel"), exclude = TRUE)

# Print the amount of memory freed by removing objects
freeMem(echo = TRUE)

}
\seealso{
\code{\link{rm}}, \code{\link{gc}}
}
