% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstring.R
\name{fstring}
\alias{fstring}
\title{String Formatting Function}
\usage{
fstring(pat, vect = NULL)
}
\arguments{
\item{pat}{A character string with placeholders in the form "{i}", where i is the index of the value to insert or the name of a variable in the calling environment.}

\item{vect}{A character vector of values to insert into the placeholders. If \code{NULL}, the function looks for variables in the calling environment with the same name as the placeholders.}
}
\value{
A character string with the placeholders replaced by values from \code{vect} or the calling environment.
}
\description{
This function replaces placeholders in a character string with values from a vector or from variables in the calling environment.
}
\examples{
fstring("The value of x is {1} and the value of y is {2}", c(3, 4))
fstring("The value of x is {x} and the value of y is {y}")
}
\keyword{character}
