% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%-\%}
\alias{\%-\%}
\title{Subtract Corresponding Elements of Two Vectors or Lists}
\usage{
x \%-\% y
}
\arguments{
\item{x}{A numeric vector or list.}

\item{y}{A numeric vector or list of the same length as `x`.}
}
\value{
A vector or list of the same length as `x` and `y`, with each element being the difference of corresponding elements of `x` and `y`.
}
\description{
This operator subtracts corresponding elements of two vectors or lists. If both elements are `NA`, the result is `NA`, otherwise `NA` values are removed before subtraction.
}
\examples{
c(1, 2, 3) \%-\% c(4, 5, 6)  # Returns c(-3, -3, -3)
c(1, NA, 3) \%-\% c(NA, 2, NA)  # Returns c(1, -2, 3)
}
