% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%+\%}
\alias{\%+\%}
\title{Add Corresponding Elements of Two Vectors or Lists}
\usage{
x \%+\% y
}
\arguments{
\item{x}{A numeric vector or list.}

\item{y}{A numeric vector or list of the same length as `x`.}
}
\value{
A vector or list of the same length as `x` and `y`, with each element being the sum of corresponding elements of `x` and `y`.
}
\description{
This operator adds corresponding elements of two vectors or lists, removing NAs.
}
\examples{
c(1, 2, 3) \%+\% c(4, 5, NA)  # Returns c(5, 7, 3)
}
