% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsplit.R
\name{hsplit}
\alias{hsplit}
\title{Horizontal Split}
\usage{
hsplit(df, cnum, multisplit = F)
}
\arguments{
\item{df}{Object to split}

\item{cnum}{number of column to split. If `multisplit` `TRUE` it is the number of columns per resulting dataframe (the last `data.frame` is the remaining columns).}

\item{multisplit}{if `FALSE` split `df` in two. If `TRUE` split `df` in multiple data.frames with `cnum` column number, with the last one being the remainder columns.}
}
\description{
Splits `data.frame`, `tibble`, or `matrix` by column number.
}
\examples{
dataframe <- data.frame(a = 1:5, b = 1:5, c = 1:5, d = 1:5, e = 1:5, f = 1:5)

## split in two
hsplit(dataframe, cnum = 3)

## split in multiple
hsplit(dataframe, cnum = 2, multisplit = T)
}
