% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pak.R
\name{load_daily_packages}
\alias{load_daily_packages}
\title{Load Daily Packages}
\usage{
load_daily_packages()
}
\value{
Loads the specified packages into the R environment.
}
\description{
This function utilizes the `pacman` package to load a set of commonly used packages for daily data analysis tasks.
The following packages are loaded:
}
\details{
\describe{
  \item{rio}{A package for streamlined data import and export.}
  \item{tidyverse}{A collection of packages for data manipulation, visualization, and more, including `dplyr`, `ggplot2`, `tidyr`, and others.}
  \item{DescTools}{A package providing a variety of descriptive statistics tools and tests.}
  \item{janitor}{A package for data cleaning and examination, offering functions like `clean_names` and `remove_empty`.}
  \item{flextable}{A package for creating and formatting tables for reporting.}
  \item{openxlsx}{A package for reading, writing, and manipulating Excel files without the need for external dependencies.}
}
}
\examples{
load_daily_packages()
}
