% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_allocation.R
\name{powerAlloc}
\alias{powerAlloc}
\title{Calculate Power Allocation}
\usage{
powerAlloc(target = NULL, V1 = NULL, power = 0.5, V2 = NULL, minAlloc = NULL)
}
\arguments{
\item{target}{Numeric. The total allocation target.}

\item{V1}{Numeric vector. Primary allocation vector.}

\item{power}{Numeric. The power to which each element of `V1` is raised, default is 0.5.}

\item{V2}{Numeric vector. Optional secondary allocation vector. Must have the same length as `V1` if provided.}

\item{minAlloc}{Numeric. Optional minimum allocation for each element of `V1`.}
}
\value{
A numeric vector of allocated values.
}
\description{
This function allocates resources (e.g., sample sizes) to different groups or strata based on a target total, using a power allocation method. Optional minimum allocations and auxiliary information can be used to guide the allocation.
}
\examples{
powerAlloc(target = 100, V1 = c(30, 20, 50), power = 0.5, V2 = c(35, 25, 40), minAlloc = 10)
}
