% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum_sample_size.R
\name{samplesize_prop}
\alias{samplesize_prop}
\title{Calculate Sample Size for Proportion}
\usage{
samplesize_prop(
  p,
  e,
  N,
  alpha = 0.05,
  deff = 1,
  rr = 0.9,
  moe_type = "relative"
)
}
\arguments{
\item{p}{Numeric. The estimated proportion (must be between 0 and 1).}

\item{e}{Numeric. The margin of error (must be between 0 and 1).}

\item{N}{Numeric. The population size (must be greater than 0).}

\item{alpha}{Numeric. The significance level, default is 0.05.}

\item{deff}{Numeric. The design effect, default is 1.}

\item{rr}{Numeric. The expected response rate, default is 0.9.}

\item{moe_type}{Character. The type of margin of error, either "relative" or "absolute", default is "relative".}
}
\value{
The calculated sample size.
}
\description{
This function calculates the required sample size for estimating a proportion in a finite population, given the desired margin of error, confidence level, design effect, and expected response rate.
}
\examples{
samplesize_prop(p = 0.5, e = 0.05, N = 10000)
}
