% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum_sample_size.R
\name{samplesize_res_prev}
\alias{samplesize_res_prev}
\title{Calculate Sample Size Based on Previous Survey's Relative Standard Error}
\usage{
samplesize_res_prev(rse_target, rse_prev, n_prev, N, rr)
}
\arguments{
\item{rse_target}{Numeric. The target relative standard error (must be greater than 0).}

\item{rse_prev}{Numeric. The relative standard error from the previous survey (must be greater than 0).}

\item{n_prev}{Numeric. The sample size from the previous survey which generated the given RSE (must be greater than 0).}

\item{N}{Numeric. The population size (must be greater than 0).}

\item{rr}{Numeric. The expected response rate (must be between 0 and 1).}
}
\value{
The calculated sample size.
}
\description{
This function calculates the required sample size to achieve a target relative standard error (RSE) for a finite population, based on the RSE and sample size from a previous survey.
}
\examples{
samplesize_res_prev(rse_target = 0.05, rse_prev = 0.1, n_prev = 500, N = 10000, rr = 0.9)
}
