% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum_sample_size.R
\name{samplesize_slovin}
\alias{samplesize_slovin}
\title{Calculate Sample Size Using Slovin's Formula}
\usage{
samplesize_slovin(d, N, deff = 1, rr = 0.9)
}
\arguments{
\item{d}{Numeric. The absolute margin of error (must be between 0 and 1).}

\item{N}{Numeric. The population size (must be greater than 0).}

\item{deff}{Numeric. The design effect, default is 1.}

\item{rr}{Numeric. The expected response rate, default is 0.9.}
}
\value{
The calculated sample size.
}
\description{
This function calculates the required sample size for a finite population using Slovin's formula, given the desired absolute margin of error, design effect, and expected response rate.
}
\examples{
samplesize_slovin(d = 0.05, N = 10000)
}
