% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsInf.R
\name{subsInf}
\alias{subsInf}
\title{Replace infinite values in a matrix with a specified value}
\usage{
subsInf(x, replace_with = NA)
}
\arguments{
\item{x}{A matrix, data frame, or tibble}

\item{replace_with}{An optional value to replace infinite values with. Defaults to \code{NA}.}
}
\value{
A modified matrix with infinite values replaced by \code{replace_with}.
}
\description{
This function takes in a matrix-like object and replaces any infinite values in the matrix with a specified value, which defaults to \code{NA}.
}
\examples{
subsInf(matrix(c(1, 2, -Inf, 4, 5, 6, Inf, Inf, 9), nrow = 3), replace_with = -999)
subsInf(matrix(c(1, 2, Inf, 4, 5, Inf), nrow = 2))

}
