% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeMem.R
\name{freeMem}
\alias{freeMem}
\title{Free Memory}
\usage{
freeMem(objs = NULL, echo = F)
}
\arguments{
\item{objs}{Character vector of the names of objects to delete. If left empty all objects will be deleted}

\item{echo}{If true, output for gc() is shown. F by default}
}
\description{
Deletes selected files from the environment
}
\examples{
freeMem(c("a", "b"), echo = T)
}
