% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsplit.R
\name{hsplit}
\alias{hsplit}
\title{Split a data frame or matrix into two or more horizontal sections}
\usage{
hsplit(df, cnum, multisplit = F)
}
\arguments{
\item{df}{A data frame, tibble, or matrix to be split.}

\item{cnum}{The column number to split the data frame or matrix on.}

\item{multisplit}{A logical value indicating whether to split the data frame
or matrix into multiple sections. The default is \code{FALSE}.}
}
\value{
If \code{multisplit} is \code{FALSE}, this function returns a list
  containing two data frames: the first containing columns 1 through \code{cnum},
  and the second containing columns (\code{cnum}+1) through the end. If
  \code{multisplit} is \code{TRUE}, this function returns a list of data frames,
  with each data frame containing the number of columns specified in \code{cnum}.
}
\description{
This function splits a data frame or matrix into two sections based on the
column number specified in the \code{cnum} argument. If \code{multisplit} is
set to \code{TRUE}, the data frame or matrix is split into multiple sections
based on the number of columns specified in \code{cnum}.
}
\examples{
# Split a data frame into two sections
hsplit(iris, 3)

# Split a matrix into two sections
hsplit(matrix(1:20, nrow = 4), 2)

# Split a data frame into multiple sections
hsplit(iris, 2, multisplit = TRUE)

}
\seealso{
\code{\link{vsplit}}
}
